+++ modules/memberFetcher.js
@@ -1,5 +1,5 @@
-module.exports.fetchMembers = async message => {
-	return await message.guild.members.fetch()
+module.exports.fetchMembers = async (message, withPresences = false) => {
+	return await message.guild.members.fetch({withPresences: withPresences})
 		.catch(err => {
 			console.error("Failed to fetch members in object resolver: " + err);
 			return message.guild.members.cache;